requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
orjson
pandas
numpy
scikit-learn
scipy
rapidfuzz
sentence-transformers
python-dotenv
httpx
structlog
loguru
streamlit
pytest
pytest-cov
ruff
mypy
pyyaml
